# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Timeful.",
        "title": "Timeful API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3002/api",
    "paths": {
...
        "/analytics/user/{email}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns a minimal public user profile (safe for unauthenticated clients)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{userId}/is-premium": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns whether the given user is a premium user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "isPremium": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.AppleCalendarAuth": {
            "type": "object"
        },
...
        "models.BufferTimeOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "models.CalendarAccount": {
            "type": "object",
            "properties": {
                "appleCalendarAuth": {
                    "$ref": "#/definitions/models.AppleCalendarAuth"
                },
                "calendarType": {
                    "$ref": "#/definitions/models.CalendarType"
                },
                "email": {
                    "description": "Email is required for all calendar accounts",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "icsCalendarAuth": {
                    "$ref": "#/definitions/models.ICSCalendarAuth"
                },
                "oAuth2CalendarAuth": {
                    "$ref": "#/definitions/models.OAuth2CalendarAuth"
                },
                "picture": {
                    "type": "string"
                },
                "subCalendars": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SubCalendar"
                    }
                }
            }
        },
...
        "models.CalendarOptions": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "$ref": "#/definitions/models.BufferTimeOptions"
                },
                "workingHours": {
                    "$ref": "#/definitions/models.WorkingHoursOptions"
                }
            }
        },
        "models.CalendarType": {
            "type": "string",
            "enum": [
                "apple",
                "google",
                "outlook",
                "ics"
            ],
            "x-enum-varnames": [
                "AppleCalendarType",
                "GoogleCalendarType",
                "OutlookCalendarType",
                "ICSCalendarType"
            ]
        },
...
        "models.ICSCalendarAuth": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                }
            }
        },
...
        "models.OAuth2CalendarAuth": {
            "type": "object"
        },
...
        "models.SubCalendar": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Profile info",
                    "type": "string"
                },
                "calendarAccounts": {
                    "description": "CalendarAccounts is a mapping from {`email_CALENDARTYPE` =\u003e CalendarAccount} that contains all the\nadditional accounts the user wants to see google calendar events for",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.CalendarAccount"
                    }
                },
                "calendarOptions": {
                    "description": "Calendar options",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CalendarOptions"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasCustomName": {
                    "description": "Whether the user has set a custom name for themselves, i.e. don't change their name when they sign in",
                    "type": "boolean"
                },
                "isPremium": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numEventsCreated": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "primaryAccountKey": {
                    "description": "The calendarAccountKey of the account the user first signed in with",
                    "type": "string"
                },
                "stripeCustomerId": {
                    "description": "Stripe customer ID",
                    "type": "string"
                },
                "timezoneOffset": {
                    "type": "integer"
                }
            }
        },
        "models.WorkingHoursOptions": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "endTime": {
                    "type": "number"
                },
                "startTime": {
                    "type": "number"
                }
            }
        },
...
    }
}